==> BLOC_1/src/02-Weather-villes.ipynb <==
# 📍 Récupération des informations Météo pour chaque villes via l’API OpenWeatherMap

Ce notebook utilise l'API [Openweathermap](https://home.openweathermap.org/) pour récupérer les informations météo de chaque villes.
Il exporte ensuite les résultats dans un nouveau fichier CSV (reprenant les coordonnées GPS déjà obtenues).

import pandas as pd
import requests
import time

# Clé API personnelle OpenWeatherMap
API_KEY = '7b6d3304dbb89b1c5e83fcbd8f1ff269'

# Chargement du fichier CSV
df = pd.read_csv(r"C:\Users\david\Desktop\FULLSTACK-Projects\FULLSTACK-Projects\BLOC_1\Data\coordonnees_villes.csv")

# Fonction pour récupérer la météo actuelle et les prévisions 5 jours
def get_weather_forecast(lat, lon):
    url = f"https://api.openweathermap.org/data/2.5/forecast"
    params = {
        'lat': lat,
        'lon': lon,
        'appid': API_KEY,
        'units': 'metric',
        'lang': 'fr'
    }
    response = requests.get(url, params=params)
    if response.status_code == 200:
        return response.json()
    else:
        print(f"Erreur API: {response.status_code}")
        return None

# Structure pour stocker les résultats
rows = []

# Boucle sur chaque ligne du CSV
# for i, row in df.iterrows():
#     lat = row['lat']
#     lon = row['lon']
#     ville = row.get('ville', f'Ville_{i}')

for i, row in df.iterrows():
    lat = row['lat']
    lon = row['lon']
    ville = row.get('ville', f'Ville_{i}')
    ville_id = row['ville_id']  


    print(f"🔄 Récupération météo pour {ville}...")

    data = get_weather_forecast(lat, lon)
    if data is None:
        continue

    for entry in data['list']:
        date_time = entry['dt_txt']
        temperature = entry['main']['temp']
        humidity = entry['main']['humidity']
        weather = entry['weather'][0]['description']

        rows.append({
            'ville_id' : ville_id,
            'ville': ville,
            'lat': lat,
            'lon': lon,
            'datetime': date_time,
            'temperature': temperature,
            'humidity': humidity,
            'weather': weather
        })

    time.sleep(1)  # éviter d’être bloqué par l’API gratuite

# Création du DataFrame enrichi
df_meteo = pd.DataFrame(rows)

# Sauvegarde dans un nouveau fichier CSV
df_meteo.to_csv("coordonnees-weather-villes.csv", index=False, encoding='utf-8-sig')

print("✅ Données météo sauvegardées dans coordonnees-weather-villes.csv")
